requirements.txt
# Framework web
fastapi
uvicorn[standard]

# Manipulation des requêtes HTTP
requests

# Gestion des modèles de données
pydantic

# Variables d'environnement
python-dotenv

# NLP et embeddings
spacy
nltk
sentence-transformers

# Base vectorielle
qdrant-client

# OCR et PDF
pdf2image
pytesseract
poppler-utils  # Nécessaire pour `pdf2image` sous Linux
flair
pyspellchecker
camelot-py